Refactored to make this compatible with legacy rosetta types
