+++ src/main/java/com/regnosys/rosetta/common/serialisation/ObjectMapperCreator.java
@@ -1,0 +1,9 @@
+package com.regnosys.rosetta.common.serialisation;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public interface ObjectMapperCreator {
+
+    ObjectMapper create();
+
+}
