+++ src/main/java/com/regnosys/rosetta/common/serialisation/RosettaSerialiserException.java
@@ -1,0 +1,9 @@
+package com.regnosys.rosetta.common.serialisation;
+
+public class RosettaSerialiserException extends RuntimeException {
+    private static final long serialVersionUID = 1L;
+
+    public RosettaSerialiserException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
