+++ src/main/java/com/regnosys/rosetta/common/serialisation/mixin/ReferenceWithMetaMixIn.java
@@ -1,0 +1,7 @@
+package com.regnosys.rosetta.common.serialisation.mixin;
+
+import com.fasterxml.jackson.annotation.JsonFilter;
+
+@JsonFilter("ReferenceFilter")
+public interface ReferenceWithMetaMixIn {
+}
