+++ src/main/java/com/regnosys/rosetta/common/serialisation/mixin/legacy/LegacyReferenceMixIn.java
@@ -1,0 +1,8 @@
+package com.regnosys.rosetta.common.serialisation.mixin.legacy;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public interface LegacyReferenceMixIn {
+    @JsonProperty("value")
+    String getReference();
+}
